setting an agents labels after is it created will not apply instantly
This causes some flakes as the labels will not be re-evaluated until a
periodic `trimLabels` in the future or a slave being added removed that
has the same labels as the agent that had setLabels called on it.

Additionally we do not need to wait for the agent to come online as the
jobs will wait anyway - this allows us to speed up the tests as we can
continue to do some work whilst the agent is connecting.