Require Jenkins 2.401.3 or newer (#237)
* Require Jenkins 2.401.3 or newer

https://stats.jenkins.io/pluginversions/declarative-pipeline-migration-assistant.html shows:

* 88% of the installations of the most recent release are already running
  Jenkins 2.401.3 or newer.  The most recent release was 4 months ago.
  The most recent release is installed on 460 / 1388 total installations.

* 48% of the installations of the next most recent release are already
  running Jenkins 2.401.3 or newer.  That release was 6 months ago.
  That release is installed on 104 / 1388 total installations.

Updating to require Jenkins 2.401.3 or newer allows us to rely on the
plugin bill of materials for two more version declarations:

* GitHub plugin
* Ansicolor plugin

Updating to require Jenkins 2.401.3 replaces the following two pull requests:

* #233 https://github.com/jenkinsci/declarative-pipeline-migration-assistant-plugin/pull/233
* #234 https://github.com/jenkinsci/declarative-pipeline-migration-assistant-plugin/pull/234

* Test with JIRA plugin 3.12 to avoid outdated API jar dependency

The Artifactory bandwidth reduction project described at
https://github.com/jenkins-infra/helpdesk/issues/3842 has shown that some
outdated JIRA API jar files are only available from the jcenter cache.
Let's test with a newer version of the Jira plugin to avoid that older
JIRA API jar file dependency.

JIRA plugin 3.12 fixes a security vulnerability. It was released 4
weeks ago.

* Depend on build timeout plugin 1.31

Build timeout plugin 1.31 includes the removal of Prototype.js that is
required for Jenkins 2.426.1 and newer.  It was released 7 months ago
and is already installed on 40% of all build timeout plugin installations.

* Use parent pom 4.76

* Test with Java 21 and Java 17

Java 21 was released Sep 19, 2023. Full support for Java 21 was
announced in November 2023.

The acceptance test harness and plugin bill of materials tests are
already passing with Java 21. This is a further step to improve plugin
readiness for use with Java 21 and for development with Java 21.

The change intentionally tests only two Java configurations, Java 17
and Java 21 because we believe that the risk of a regression that only
affects Java 11 is low. We generate Java 11 byte code with the Java 17
and the Java 21 builds, so we're already testing Java 11 byte code.

Java 11 is also tested for this plugin in the plugin bill of materials
each week.